Some linting cleanup and a minor refactor for executor location
