+++ hell_snake.py
@@ -1,7 +1,6 @@
 #!.venv/bin/python3
 import constants
 import sys
-from src.settings import SettingsManager
 from src.controller import Controller
 from src.model import Model
 
@@ -16,7 +15,7 @@ def main():
         view = PyQT5View(controller)
     else:
         # If the view framework is not supported or recognized, print an error and exit.
-        print(f"Error: Unsupported view framework '{settings.view_framework}'.", file=sys.stderr)
+        print(f"Error: Unsupported view framework '{model.settingsManager.view_framework}'.", file=sys.stderr)
         print("Please configure a supported view framework in the settings.", file=sys.stderr)
         sys.exit(1)
