+++ src/view/pyqt5/edit_config_dialog.py
@@ -2,8 +2,7 @@ import constants
 from PyQt5.QtWidgets import QDialog, QTabWidget, QVBoxLayout, QLabel, QGridLayout, QPushButton, QWidget, QFrame
 from PyQt5.QtGui import QFont, QIcon
 from PyQt5.QtCore import Qt
-from src.settings import SettingsManager
-from src.view.pyqt5.util import show_capture_key_dialog, DropdownDialog, NumberInputDialog, PyQT5Settings
+from src.view.pyqt5.util import show_capture_key_dialog, DropdownDialog, NumberInputDialog
 
 class EditConfigDialog(QDialog):
     def __init__(self, controller):
