+++ src/view/pyqt5/edit_loadout_dialog.py
@@ -5,7 +5,7 @@ from PyQt5.QtCore import QSize, Qt
 from PyQt5.QtSvg import QSvgWidget
 from copy import deepcopy
 from src.view.pyqt5.filter_dialog import FilteredListDialog
-from src.view.pyqt5.util import show_capture_key_dialog, PyQT5Settings
+from src.view.pyqt5.util import show_capture_key_dialog
 
 class EditLoadoutDialog(QDialog):
     def __init__(self, controller):
