+++ src/view/pyqt5/util.py
@@ -1,7 +1,6 @@
 from PyQt5.QtWidgets import QMessageBox, QDialog, QVBoxLayout, QLabel, QComboBox, QPushButton, QHBoxLayout, QLineEdit
 from PyQt5.QtGui import QIntValidator
 from PyQt5.QtCore import QObject, pyqtSignal, QThread
-from src.settings import SettingsManager
 
 KEY_ALWAYS_ON_TOP = "pyqt5_alwaysOnTop"
 KEY_ALWAYS_ON_TOP_DEFAULT = False
